requirements.txt
Flask
gunicorn
google-cloud-storage
google-generativeai
pillow

